+++ R/ref-grid.R
@@ -263,14 +263,15 @@
 #'   \samp{transform = "log", type = "response"}. See also the help page for 
 #'   \code{\link{regrid}}.
 #'
-#' @section Side effect: The most recent result of \code{ref_grid}, whether
+#' @section Optional side effect: If the \code{save.ref_grid} option is set to
+#'   \code{TRUE} (see \code{\link{emm_options}}),
+#'   The most recent result of \code{ref_grid}, whether
 #'   called directly or indirectly via \code{\link{emmeans}},
 #'   \code{\link{emtrends}}, or some other function that calls one of these, is
 #'   saved in the user's environment as \code{.Last.ref_grid}. This facilitates
 #'   checking what reference grid was used, or reusing the same reference grid
-#'   for further calculations. This automatic saving is enabled by default, but
-#'   may be disabled via \samp{emm_options(save.ref_grid = FALSE)}, and
-#'   re-enabled by specifying \code{TRUE}.
+#'   for further calculations. This automatic saving is disabled by default, but
+#'   may be enabled via \samp{emm_options(save.ref_grid = TRUE)}.
 #'
 #' @return An object of the S4 class \code{"emmGrid"} (see
 #'   \code{\link{emmGrid-class}}). These objects encapsulate everything needed
@@ -301,7 +302,6 @@
 #' @examples
 #' fiber.lm <- lm(strength ~ machine*diameter, data = fiber)
 #' ref_grid(fiber.lm)
-#' summary(.Last.ref_grid)
 #'
 #' ref_grid(fiber.lm, at = list(diameter = c(15, 25)))
 #'